refactored the eventChildFetch & eventParentFetch
created full blown schema mapping for either movie results or event results. standardized the method in which the filterType is used to determine what gets called in the eventParent and conversely also the child. added encoding types for the input/output streams on the child. moved preliminary request’s out into drivers that are required by the module on instantiation.