refactored yelpChildFetch and yelpParentFetch
Changed some of the variables to reflect a more hierarchical understanding of how the driver relates to the child process. Added a fundamental understanding and documentation of the YelpParent (still needs more work). like previous commit, added the process pid of the child process for monitoring purposes. Added the beaches category to the yelp list.