move mapShowtimes around inside the gracenote driver
tried this inside/tried this outside. I think this mapShowtimes function will have to live inside the main requesting function of the driver, because it share the location data and since i can’t exactly pass that off when function is async.mapping. there has to be an easier way to do this… :sad_face: